refactor : CarNameValidator를 util에서 domain으로 이동
