+++ src/main/java/racingcar/domain/CarNameValidator.java
@@ -1,4 +1,4 @@
-package racingcar.utils;
+package racingcar.domain;
 
 import java.util.Arrays;
 import java.util.HashSet;
